==> 2.ipynb <==
import pandas as pd
import numpy as np

credits = pd.read_csv('credits.csv')

credits

meta = pd.read_csv('movies_metadata.csv')

meta.columns

meta['release_date'] = pd.to_datetime(meta['release_date'], errors='coerce')

meta['year'] = meta['release_date'].dt.year

meta['year'].value_counts().sort_index()

new_meta = meta.loc[meta.year == 2017,['genres', 'id', 'title', 'year']]

new_meta['id'] = new_meta['id'].astype(int)

df = pd.merge(new_meta, credits, on='id')

pd.set_option('display.max_colwidth', 75)
df

import ast
df['genres'] = df['genres'].map(lambda x: ast.literal_eval(x))
df['cast'] = df['cast'].map(lambda x: ast.literal_eval(x))
df['crew'] = df['crew'].map(lambda x: ast.literal_eval(x))

def make_genreslist(x):
    gen = []
    st = ' '
    for i in x:
        if i.get('name') == 'Science Fiction':
            scifi = 'Sci-Fi'
            gen.append(scifi)
        else:
            gen.append(i.get('name'))
    if gen == []:
        return np.NaN
    else:
        return (st.join(gen))

df['genres_list'] = df['genres'].map(lambda x: make_genreslist(x))

df['genres_list']

def get_actor1(x):
    casts = []
    for i in x:
        casts.append(i.get('name'))
        if casts == []:
            return np.NaN
        else:
            return (casts[0])

df['actor_1_name'] = df['cast'].map(lambda x: get_actor1(x))

def get_actor2(x):
    casts = []
    for i in x:
        casts.append(i.get('name'))
    if casts == [] or len(casts)<=1:
        return np.NaN
    else:
        return (casts[1])

df['actor_2_name'] = df['cast'].map(lambda x: get_actor2(x))

df['actor_2_name']

def get_actor3(x):
    casts = []
    for i in x:
        casts.append(i.get('name'))
    if casts == [] or len(casts)<=2:
        return np.NaN
    else:
        return (casts[2])

df['actor_3_name'] = df['cast'].map(lambda x: get_actor3(x))

def get_directors(x):
    dt = []
    st = " "
    for i in x:
        if i.get('job') == 'Director':
            dt.append(i.get('name'))
    if dt == []:
        return np.NaN
    else:
        return (st.join(dt))

df['director_name'] = df['crew'].map(lambda x: get_directors(x))

movie = df.loc[:,['director_name','actor_1_name','actor_2_name','actor_3_name','genres_list','title']]

movie

movie.isna().sum()

movie = movie.dropna(how='any')

movie.isna().sum()

movie = movie.rename(columns={'genres_list': 'genres'})
movie = movie.rename(columns={'title': 'movie_title'})

movie['movie_title'] = movie['movie_title'].str.lower()

movie['comb'] = movie['actor_1_name'] + ' ' + movie['actor_2_name'] + ' '+ movie['actor_3_name'] + ' '+ movie['director_name'] +' ' + movie['genres']

movie

old = pd.read_csv('data.csv')

old

old['comb'] = old['actor_1_name'] +' ' +old['actor_2_name']+' '+old['actor_3_name']+' '+old['director_name']+' '+old['genres']

old

new = old.append(movie)

new

new.drop_duplicates(subset='movie_title', keep='last', inplace=True)

new

new.to_csv('new_data.csv', index=False)
