==> 5.ipynb <==
import pandas as pd
import numpy as np
import nltk
from nltk.corpus import stopwords
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.model_selection import train_test_split
from sklearn import naive_bayes
from sklearn.metrics import roc_auc_score,accuracy_score
import pickle

nltk.download("stopwords")

df = pd.read_csv("reviews.txt",sep='\t', names=['Reviews', 'Comments'])

df

stopset = set(stopwords.words('english'))

vectorizer = TfidfVectorizer(use_idf = True, lowercase=True, strip_accents='ascii', stop_words=stopset)

X= vectorizer.fit_transform(df.Comments)
y = df.Reviews
pickle.dump(vectorizer, open('tranform1.pkl','wb'))

X_train, X_test, y_train, y_test = train_test_split(X,y, test_size=0.2, random_state=42)

clf = naive_bayes.MultinomialNB()
clf.fit(X_train, y_train)

accuracy_score(y_test, clf.predict(X_test))*100

clf = naive_bayes.MultinomialNB()
clf.fit(X,y)

accuracy_score(y_test, clf.predict(X_test))*100

filename = 'model.pkl'
pickle.dump(clf, open(filename, 'wb'))
